move search and rename the decorated index with pupiltype
