+++ DfE.GIAP.All/src/DfE.GIAP.Core/MyPupils/Application/Services/AggregatePupilsForMyPupils/DataTransferObjects/AzureIndexEntityWithPupilType.cs
@@ -1,0 +1,6 @@
+﻿using DfE.GIAP.Core.MyPupils.Domain.ValueObjects;
+
+namespace DfE.GIAP.Core.MyPupils.Application.Services.AggregatePupilsForMyPupils.DataTransferObjects;
+public sealed record AzureIndexEntityWithPupilType(
+    AzureIndexEntity SearchIndexDto,
+    PupilType PupilType);
