+++ DfE.GIAP.All/src/DfE.GIAP.Core/MyPupils/Application/Services/AggregatePupilsForMyPupils/IAggregatePupilsForMyPupilsService.cs
@@ -1,7 +1,7 @@
 ﻿using DfE.GIAP.Core.MyPupils.Domain.Entities;
 using DfE.GIAP.Core.MyPupils.Domain.ValueObjects;
 
-namespace DfE.GIAP.Core.MyPupils.Application.UseCases.GetMyPupils.Services.AggregatePupilsForMyPupils;
+namespace DfE.GIAP.Core.MyPupils.Application.Services.AggregatePupilsForMyPupils;
 public interface IAggregatePupilsForMyPupilsApplicationService
 {
     Task<IEnumerable<Pupil>> GetPupilsAsync(UniquePupilNumbers uniquePupilNumbers);
