+++ DfE.GIAP.All/src/DfE.GIAP.Core/MyPupils/Application/Services/AggregatePupilsForMyPupils/Mapper/MapDecoratedSearchIndexDtoToPupilMapper.cs
@@ -1,12 +1,11 @@
-﻿using DfE.GIAP.Core.Common.CrossCutting;
-using DfE.GIAP.Core.MyPupils.Application.UseCases.GetMyPupils.Services.AggregatePupilsForMyPupils.DataTransferObjects;
+﻿using DfE.GIAP.Core.MyPupils.Application.Services.AggregatePupilsForMyPupils.DataTransferObjects;
 using DfE.GIAP.Core.MyPupils.Domain.Entities;
 using DfE.GIAP.Core.MyPupils.Domain.ValueObjects;
 
-namespace DfE.GIAP.Core.MyPupils.Application.UseCases.GetMyPupils.Services.AggregatePupilsForMyPupils.Mapper;
-internal sealed class MapDecoratedSearchIndexDtoToPupilMapper : IMapper<DecoratedSearchIndexDto, Pupil>
+namespace DfE.GIAP.Core.MyPupils.Application.Services.AggregatePupilsForMyPupils.Mapper;
+internal sealed class MapDecoratedSearchIndexDtoToPupilMapper : IMapper<AzureIndexEntityWithPupilType, Pupil>
 {
-    public Pupil Map(DecoratedSearchIndexDto input) =>
+    public Pupil Map(AzureIndexEntityWithPupilType input) =>
         new(
             identifier: new UniquePupilNumber(input.SearchIndexDto.UPN),
             pupilType: input.PupilType,
