--- DfE.GIAP.All/src/DfE.GIAP.Core/MyPupils/Application/UseCases/GetMyPupils/Services/AggregatePupilsForMyPupils/DataTransferObjects/DecoratedSearchIndexDto.cs
@@ -1,7 +1,0 @@
-﻿using DfE.GIAP.Core.MyPupils.Application.UseCases.GetMyPupils.Services.AggregatePupilsForMyPupils.Dto;
-using DfE.GIAP.Core.MyPupils.Domain.ValueObjects;
-
-namespace DfE.GIAP.Core.MyPupils.Application.UseCases.GetMyPupils.Services.AggregatePupilsForMyPupils.DataTransferObjects;
-public sealed record DecoratedSearchIndexDto(
-    AzureIndexEntity SearchIndexDto,
-    PupilType PupilType);
