+++ DfE.GIAP.All/tests/DfE.GIAP.Core.UnitTests/MyPupils/Application/Services/TempAggregatePupilsForMyPupilsServiceTests.cs
@@ -2,9 +2,8 @@
 using DfE.GIAP.Core.Common.CrossCutting;
 using DfE.GIAP.Core.MyPupils.Application.Extensions;
 using DfE.GIAP.Core.MyPupils.Application.Search.Provider;
-using DfE.GIAP.Core.MyPupils.Application.UseCases.GetMyPupils.Services.AggregatePupilsForMyPupils;
-using DfE.GIAP.Core.MyPupils.Application.UseCases.GetMyPupils.Services.AggregatePupilsForMyPupils.DataTransferObjects;
-using DfE.GIAP.Core.MyPupils.Application.UseCases.GetMyPupils.Services.AggregatePupilsForMyPupils.Dto;
+using DfE.GIAP.Core.MyPupils.Application.Services.AggregatePupilsForMyPupils;
+using DfE.GIAP.Core.MyPupils.Application.Services.AggregatePupilsForMyPupils.DataTransferObjects;
 using DfE.GIAP.Core.MyPupils.Domain.Entities;
 using DfE.GIAP.Core.MyPupils.Domain.ValueObjects;
 using DfE.GIAP.Core.UnitTests.MyPupils.TestDoubles;
@@ -19,8 +18,8 @@ public sealed class TempAggregatePupilsForMyPupilsServiceTests
     {
         // Arrange
         Mock<ISearchClientProvider> searchClientProviderMock = SearchClientProviderTestDoubles.Default();
-        Mock<IMapper<DecoratedSearchIndexDto, Pupil>> mapper = MapperTestDoubles.MockFor<DecoratedSearchIndexDto, Pupil>();
-        TempAggregatePupilsForMyPupilsApplicationService service = new(searchClientProviderMock.Object, mapper.Object);
+        Mock<IMapper<AzureIndexEntityWithPupilType, Pupil>> mapper = MapperTestDoubles.MockFor<AzureIndexEntityWithPupilType, Pupil>();
+        AggregatePupilsForMyPupilsApplicationService service = new(searchClientProviderMock.Object, mapper.Object);
 
         // Act
         UniquePupilNumbers uniquePupilNumbers = UniquePupilNumbers.Create([]);
@@ -35,8 +34,8 @@ public sealed class TempAggregatePupilsForMyPupilsServiceTests
     {
         // Arrange
         Mock<ISearchClientProvider> searchClientProviderMock = new();
-        Mock<IMapper<DecoratedSearchIndexDto, Pupil>> mapper = MapperTestDoubles.MockFor<DecoratedSearchIndexDto, Pupil>();
-        TempAggregatePupilsForMyPupilsApplicationService service = new(searchClientProviderMock.Object, mapper.Object);
+        Mock<IMapper<AzureIndexEntityWithPupilType, Pupil>> mapper = MapperTestDoubles.MockFor<AzureIndexEntityWithPupilType, Pupil>();
+        AggregatePupilsForMyPupilsApplicationService service = new(searchClientProviderMock.Object, mapper.Object);
         IEnumerable<UniquePupilNumber> generatedUpns = UniquePupilNumberTestDoubles.Generate(4001);
         UniquePupilNumbers uniquePupilNumbers = UniquePupilNumbers.Create(generatedUpns);
 
@@ -59,8 +58,8 @@ public sealed class TempAggregatePupilsForMyPupilsServiceTests
 
         Mock<ISearchClientProvider> searchClientProviderMock = SearchClientProviderTestDoubles.MockFor(searchClientKeyToResponseData);
 
-        Mock<IMapper<DecoratedSearchIndexDto, Pupil>> mapper = MapperTestDoubles.MockFor<DecoratedSearchIndexDto, Pupil>();
-        TempAggregatePupilsForMyPupilsApplicationService service = new(searchClientProviderMock.Object, mapper.Object);
+        Mock<IMapper<AzureIndexEntityWithPupilType, Pupil>> mapper = MapperTestDoubles.MockFor<AzureIndexEntityWithPupilType, Pupil>();
+        AggregatePupilsForMyPupilsApplicationService service = new(searchClientProviderMock.Object, mapper.Object);
 
         // Act
         IEnumerable<Pupil> result = await service.GetPupilsAsync(uniquePupilNumbers);
@@ -88,8 +87,8 @@ public sealed class TempAggregatePupilsForMyPupilsServiceTests
 
         Mock<ISearchClientProvider> searchClientProviderMock = SearchClientProviderTestDoubles.MockFor(searchClientKeyToResponseData);
 
-        Mock<IMapper<DecoratedSearchIndexDto, Pupil>> mapper = MapperTestDoubles.MockFor<DecoratedSearchIndexDto, Pupil>();
-        TempAggregatePupilsForMyPupilsApplicationService service = new(searchClientProviderMock.Object, mapper.Object);
+        Mock<IMapper<AzureIndexEntityWithPupilType, Pupil>> mapper = MapperTestDoubles.MockFor<AzureIndexEntityWithPupilType, Pupil>();
+        AggregatePupilsForMyPupilsApplicationService service = new(searchClientProviderMock.Object, mapper.Object);
 
         // Act
         IEnumerable<Pupil> result = await service.GetPupilsAsync(uniquePupilNumbers);
@@ -115,8 +114,8 @@ public sealed class TempAggregatePupilsForMyPupilsServiceTests
 
         Mock<ISearchClientProvider> searchClientProviderMock = SearchClientProviderTestDoubles.MockFor(searchClientKeyToResponseData);
 
-        Mock<IMapper<DecoratedSearchIndexDto, Pupil>> mapper = MapperTestDoubles.MockFor<DecoratedSearchIndexDto, Pupil>();
-        TempAggregatePupilsForMyPupilsApplicationService service = new(searchClientProviderMock.Object, mapper.Object);
+        Mock<IMapper<AzureIndexEntityWithPupilType, Pupil>> mapper = MapperTestDoubles.MockFor<AzureIndexEntityWithPupilType, Pupil>();
+        AggregatePupilsForMyPupilsApplicationService service = new(searchClientProviderMock.Object, mapper.Object);
 
         List<UniquePupilNumber> requestUpns =
             npdResults.Concat(ppResults)
