+++ DfE.GIAP.All/tests/DfE.GIAP.Core.UnitTests/MyPupils/TestDoubles/AggregatePupilsForMyPupilsServiceTestDoubles.cs
@@ -1,4 +1,4 @@
-﻿using DfE.GIAP.Core.MyPupils.Application.UseCases.GetMyPupils.Services.AggregatePupilsForMyPupils;
+﻿using DfE.GIAP.Core.MyPupils.Application.Services.AggregatePupilsForMyPupils;
 using DfE.GIAP.Core.MyPupils.Domain.Entities;
 using DfE.GIAP.Core.MyPupils.Domain.ValueObjects;
