+++ DfE.GIAP.All/tests/DfE.GIAP.Core.UnitTests/MyPupils/TestDoubles/AzureIndexEntityDtosTestDoubles.cs
@@ -1,5 +1,5 @@
 ﻿using Bogus;
-using DfE.GIAP.Core.MyPupils.Application.UseCases.GetMyPupils.Services.AggregatePupilsForMyPupils.Dto;
+using DfE.GIAP.Core.MyPupils.Application.Services.AggregatePupilsForMyPupils.DataTransferObjects;
 using DfE.GIAP.Core.MyPupils.Domain.ValueObjects;
 using DfE.GIAP.SharedTests.TestDoubles;
